aws-sqs/image-converter/src/requirements.txt
Pillow==10.0.1
boto3==1.34.0
botocore==1.34.0
